[fix] Ensure background tasks are executed on transaction commit
Many tasks were being called without transaction.on_commit,
which does not guarantee the tasks will be executed after
changes have been committed to the database, which in turn
leads to inconsistent results.
