+++ openwisp_monitoring/device/models.py
@@ -1,8 +1,8 @@
 from django.contrib.contenttypes.fields import GenericRelation
 from swapper import get_model_name, swappable_setting
 
+from ..utils import load_model_patched
 from .base.models import AbstractDeviceData, AbstractDeviceMonitoring
-from .utils import load_model_patched
 
 BaseDevice = load_model_patched('config', 'Device', require_ready=False)
