+++ openwisp_monitoring/device/utils.py
@@ -1,6 +1,3 @@
-from django.apps import apps
-from swapper import is_swapped, split
-
 from ..db import timeseries_db
 from . import settings as app_settings
 
@@ -17,13 +14,3 @@ def manage_short_retention_policy():
     """
     duration = app_settings.SHORT_RETENTION_POLICY
     timeseries_db.create_or_alter_retention_policy(SHORT_RP, duration)
-
-
-def load_model_patched(app_label, model, require_ready=True):
-    """
-    TODO: remove if https://github.com/wq/django-swappable-models/pull/23 gets merged
-    """
-    swapped = is_swapped(app_label, model)
-    if swapped:
-        app_label, model = split(swapped)
-    return apps.get_model(app_label, model, require_ready=require_ready)
